+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * [Dio] Ref: Replace FailedRequestAdapter with FailedRequestInterceptor (#728)
 * Fix: Add missing return values - dart analyzer (#742)
 * Bump: Kotlin plugin to 1.5.31 (#763)
+* Feature: Add `DioEventProcessor` which improves DioError crash reports (#718)
 * Fix: Do not report duplicated packages and integrations (#760)
 
 # 6.3.0